requirements.txt
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
